move hijack loggers to utils
