+++ solarforecastarbiter/reports/tests/test_reports_main.py
@@ -95,53 +95,6 @@ def test_infer_timezone_agg(report_objects, single_forecast_aggregate,
     assert main.infer_timezone(rp) == 'America/Denver'
 
 
-def test_listhandler():
-    logger = logging.getLogger('testlisthandler')
-    handler = main.ListHandler()
-    logger.addHandler(handler)
-    logger.setLevel('DEBUG')
-    logger.warning('Test it')
-    logger.debug('What?')
-    out = handler.export_records()
-    assert len(out) == 1
-    assert out[0].message == 'Test it'
-    assert len(handler.export_records(logging.DEBUG)) == 2
-
-
-def test_listhandler_recreate():
-    logger = logging.getLogger('testlisthandler')
-    handler = main.ListHandler()
-    logger.addHandler(handler)
-    logger.setLevel('DEBUG')
-    logger.warning('Test it')
-    logger.debug('What?')
-    out = handler.export_records()
-    assert len(out) == 1
-    assert out[0].message == 'Test it'
-    assert len(handler.export_records(logging.DEBUG)) == 2
-
-    l2 = logging.getLogger('testlist2')
-    h2 = main.ListHandler()
-    l2.addHandler(h2)
-    l2.error('Second fail')
-    out = h2.export_records()
-    assert len(out) == 1
-    assert out[0].message == 'Second fail'
-
-
-def test_hijack_loggers(mocker):
-    old_handler = mocker.MagicMock()
-    new_handler = mocker.MagicMock()
-    mocker.patch('solarforecastarbiter.reports.main.ListHandler',
-                 return_value=new_handler)
-    logger = logging.getLogger('testhijack')
-    logger.addHandler(old_handler)
-    assert logger.handlers[0] == old_handler
-    with main.hijack_loggers(['testhijack']):
-        assert logger.handlers[0] == new_handler
-    assert logger.handlers[0] == old_handler
-
-
 def test_create_raw_report_from_data(mocker, report_objects, _test_data):
     report = report_objects[0]
     data = {}
